WIP replacing polled diagnostics with event-based implementation
