+++ opc-ua-sdk/sdk-server/src/main/java/org/eclipse/milo/opcua/sdk/server/namespaces/ServerNamespace.java
@@ -11,28 +11,43 @@
 package org.eclipse.milo.opcua.sdk.server.namespaces;
 
 import java.util.List;
+import java.util.NoSuchElementException;
 
+import org.eclipse.milo.opcua.sdk.core.AccessLevel;
 import org.eclipse.milo.opcua.sdk.server.OpcUaServer;
 import org.eclipse.milo.opcua.sdk.server.api.DataItem;
 import org.eclipse.milo.opcua.sdk.server.api.ManagedNamespaceWithLifecycle;
 import org.eclipse.milo.opcua.sdk.server.api.MonitoredItem;
-import org.eclipse.milo.opcua.sdk.server.diagnostics.DiagnosticsManager;
+import org.eclipse.milo.opcua.sdk.server.diagnostics.objects.ServerDiagnosticsObject;
+import org.eclipse.milo.opcua.sdk.server.model.nodes.objects.ServerDiagnosticsTypeNode;
 import org.eclipse.milo.opcua.sdk.server.util.SubscriptionModel;
+import org.eclipse.milo.opcua.stack.core.Identifiers;
 
 public class ServerNamespace extends ManagedNamespaceWithLifecycle {
 
     private final SubscriptionModel subscriptionModel;
-    private final DiagnosticsManager diagnosticsManager;
 
     public ServerNamespace(OpcUaServer server) {
         super(server, server.getConfig().getApplicationUri());
 
         subscriptionModel = new SubscriptionModel(server, this);
-        diagnosticsManager = new DiagnosticsManager(server, getNodeFactory(), getNodeManager());
+
+        ServerDiagnosticsTypeNode serverDiagnosticsNode = (ServerDiagnosticsTypeNode) getServer()
+            .getAddressSpaceManager()
+            .getManagedNode(Identifiers.Server_ServerDiagnostics)
+            .orElseThrow(() -> new NoSuchElementException("NodeId: " + Identifiers.Server_ServerDiagnostics));
+
+        serverDiagnosticsNode.getEnabledFlagNode().setUserAccessLevel(
+            AccessLevel.toValue(AccessLevel.READ_WRITE)
+        );
+
+        ServerDiagnosticsObject serverDiagnosticsObject = new ServerDiagnosticsObject(
+            serverDiagnosticsNode, getNodeManager()
+        );
 
         getLifecycleManager().addStartupTask(() -> VendorServerInfoNodes.add(getNodeContext()));
         getLifecycleManager().addLifecycle(subscriptionModel);
-        getLifecycleManager().addLifecycle(diagnosticsManager);
+        getLifecycleManager().addLifecycle(serverDiagnosticsObject);
     }
 
     @Override
